requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli>=1.30.0
flake8
python-dotenv>=0.5.1